Improve datatype casting to SQLMesh style

+++ models/achilles/temp_models/achilles__rawData_106.sql
@@ -14,7 +14,7 @@ from
     select
       person_id,
       datediff( op.observation_period_end_date,op.observation_period_start_date
-      ) as count_value,
+      )::FLOAT as count_value,
       ROW_NUMBER() over (
         partition by op.person_id order by op.observation_period_start_date asc
       ) as rn
