+++ models/achilles/temp_models/achilles__s_tmpach_112.sql
@@ -10,7 +10,7 @@ with rawData as (
   select
     YEAR(op1.observation_period_end_date) * 100
     + MONTH(op1.observation_period_end_date) as stratum_1,
-    COUNT(distinct op1.PERSON_ID) as count_value
+    COUNT(distinct op1.PERSON_ID)::FLOAT as count_value
   from
     `@src_database`.`@src_schema_omop`.`observation_period` as op1
   group by
