+++ models/achilles/temp_models/achilles__s_tmpach_1200.sql
@@ -13,7 +13,7 @@ select
   cast(null as VARCHAR(255)) as stratum_3,
   cast(null as VARCHAR(255)) as stratum_4,
   cast(null as VARCHAR(255)) as stratum_5,
-  count(person_id) as count_value
+  count(person_id)::FLOAT as count_value
 from `@src_database`.`@src_schema_omop`.`person` as p1
 inner join `@src_database`.`@src_schema_omop`.`care_site` as cs1
   on p1.care_site_id = cs1.care_site_id
