+++ models/achilles/temp_models/achilles__s_tmpach_1819.sql
@@ -13,7 +13,7 @@ select
   CAST(NULL as VARCHAR(255)) as stratum_3,
   CAST(NULL as VARCHAR(255)) as stratum_4,
   CAST(NULL as VARCHAR(255)) as stratum_5,
-  COUNT(m.person_id) as count_value
+  COUNT(m.person_id)::FLOAT as count_value
 from
   `@src_database`.`@src_schema_omop`.`measurement` as m
 inner join
