+++ models/achilles/temp_models/achilles__s_tmpach_401.sql
@@ -13,7 +13,7 @@ select
   CAST(NULL as VARCHAR(255)) as stratum_3,
   CAST(NULL as VARCHAR(255)) as stratum_4,
   CAST(NULL as VARCHAR(255)) as stratum_5,
-  COUNT(co.person_id) as count_value
+  COUNT(co.person_id)::FLOAT as count_value
 from
   `@src_database`.`@src_schema_omop`.`condition_occurrence` as co
 inner join
